resolved final review
